requirements.txt
alembic==1.11.3
Babel==2.12.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-Admin==1.6.1
flask-babel==3.1.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flask-file-upload==0.2.1
Flask-Migrate==4.0.4
flask-sqlalchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.20
typing-extensions==4.7.1
waitress==2.1.2
werkzeug==2.3.7
WTForms==3.0.1
